packages/server/pyproject.toml
[project]
name = "handler-server"
version = "0.1.0"
description = "Handler reference agent server"
requires-python = ">=3.11"
dependencies = [
    "handler-common",
    "google-adk",
    "python-dotenv",
    "litellm",
    "uvicorn",
    "click>=8.1.0",
]

[project.scripts]
handler-server = "handler_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
